Watch logs to account for external rollover
"External rollover" occurs when two or more handlers share a lock and a
log file and one of the two handlers rolls the file. Previously,
ArchivingFileHandler addressed external rollover by comparing the
handler-local state to the global, on-disk state, executing a rollover
if the on-disk state required one, and otherwise updating the local
state to match the on-disk state. However, this comparison process
is brittle because it depends on a monotonic clock which doesn't
slew backwards, otherwise time-based rollover may temporarily split
handler outputs, violating the invariant that all handlers which
share the same file and lock should direct their output any at given
instant to the same on-disk location.